+++ CHANGELOG.md
@@ -11,6 +11,7 @@ ADDED:
 CHANGED:
 
 - javadoc extended and improved
+- moved method isEmpty and isNotEmpty to class CollectionExtensions
 
 
 Version 4.24.0