+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 4.26.0-SNAPSHOT
 -------------
 
+CHANGED:
+
+- update of parent version to 3.8.0
+
 
 Version 4.25.0
 -------------